Rename fingerprinter_manager to collector
